feat: Header컴포넌트에 서 입력창과 헤더섹션을 다른 컴포넌트로 분리
